LPS-80122 (announcements-uad) Add base implementation class for UADAnonymizer
