+++ modules/apps/collaboration/announcements/announcements-uad/src/main/java/com/liferay/announcements/uad/anonymizer/AnnouncementsEntryUADAnonymizer.java
@@ -14,20 +14,10 @@
 
 package com.liferay.announcements.uad.anonymizer;
 
-import com.liferay.announcements.kernel.model.AnnouncementsEntry;
-import com.liferay.announcements.kernel.service.AnnouncementsEntryLocalService;
 import com.liferay.announcements.uad.constants.AnnouncementsUADConstants;
-import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.user.associated.data.anonymizer.DynamicQueryUADAnonymizer;
 import com.liferay.user.associated.data.anonymizer.UADAnonymizer;
 
-import java.util.Arrays;
-import java.util.List;
-
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Noah Sherrill
@@ -38,44 +28,5 @@ import org.osgi.service.component.annotations.Reference;
 	service = UADAnonymizer.class
 )
 public class AnnouncementsEntryUADAnonymizer
-	extends DynamicQueryUADAnonymizer<AnnouncementsEntry> {
-
-	@Override
-	public void autoAnonymize(
-			AnnouncementsEntry announcementsEntry, long userId,
-			User anonymousUser)
-		throws PortalException {
-
-		announcementsEntry.setUserId(anonymousUser.getUserId());
-		announcementsEntry.setUserName(anonymousUser.getFullName());
-
-		_announcementsEntryLocalService.updateAnnouncementsEntry(
-			announcementsEntry);
-	}
-
-	@Override
-	public void delete(AnnouncementsEntry announcementsEntry) {
-		_announcementsEntryLocalService.deleteAnnouncementsEntry(
-			announcementsEntry);
-	}
-
-	@Override
-	public List<String> getNonanonymizableFieldNames() {
-		return Arrays.asList("content", "title");
-	}
-
-	@Override
-	protected ActionableDynamicQuery doGetActionableDynamicQuery() {
-		return _announcementsEntryLocalService.getActionableDynamicQuery();
-	}
-
-	@Override
-	protected String[] doGetUserIdFieldNames() {
-		return AnnouncementsUADConstants.
-			USER_ID_FIELD_NAMES_ANNOUNCEMENTS_ENTRY;
-	}
-
-	@Reference
-	private AnnouncementsEntryLocalService _announcementsEntryLocalService;
-
+	extends BaseAnnouncementsEntryUADAnonymizer {
 }
